### 2025-03-20 #7931 @samhforbes
Hi @AlidesChimin thanks for your submission. Due to the relatively small size of this repository we are going to ask the editors to check this is in scope, particularly with regards to our requirements for substantial scholarly effort.
