### 2025-03-25 #7931 @AlidesChimin
Thank you for your initial assessment. Although the SATA repository may appear modest in size, it is the result of several months of academic work grounded in the methodology of Laurence Bardin (2011) and focused on enabling replicable content analysis in qualitative research.
The software is currently being used by two research groups based in Brazil, both active in graduate-level (Master’s and PhD) programs in Geography:

- GETE – Grupo de Estudos Territoriais (Territorial Studies Group) at the Universidade Estadual de Ponta Grossa (UEPG).
- GEPES – Grupo de Pesquisa Redes de Poder, Migrações e Dinâmicas Territoriais (Research Group on Power Networks, Migrations, and Territorial Dynamics) at the Universidade Estadual do Centro-Oeste (UNICENTRO).

These groups conduct research involving textual analysis in the fields of Geography, Gender Studies, and Education. SATA has proven to be a valuable tool for filtering relevant terms, extracting bigrams for semantic network analysis (with Gephi integration), classifying gender in textual corpora, and generating lexical and sentiment statistics — all through a graphical interface that is accessible to researchers without programming experience.
I would be glad to provide examples of its application in current academic projects or further enhance the documentation and tutorials if that would help. I believe SATA fills a relevant methodological gap for Humanities and Social Science researchers seeking reproducible and open-source tools for content analysis.
