### 2025-03-26 #7931 @AlidesChimin
> Envio sinalizado para revisão editorial.

Thank you for your consideration. Although the SATA repository may appear modest in size, it reflects several months of academic development grounded in the content analysis methodology of Laurence Bardin (2011). The tool was designed to support reproducible and accessible textual analysis for researchers in the Humanities and Social Sciences.
Importantly, SATA is currently being used in active Master’s and PhD research projects in Brazil. It supports methodological workflows for qualitative content analysis in multiple dissertations and theses within graduate programs in Geography at Universidade Estadual de Ponta Grossa (UEPG) and Universidade Estadual do Centro-Oeste (UNICENTRO). Examples include:

Charlley Freitas Campos (MSc, UEPG) – Territorial dynamics and resistance mechanisms of effeminate gay men.
Paulo Ernesto Lopes Rickli (MSc, UNICENTRO) – Political contestation through punk music in geographic space.
Keila Miranda Tachevski (PhD, UNICENTRO) – Intersectional analysis of women's participation in politics (gender, ethnicity, and religion).
Paulo Henrique Carneiro (PhD, UEPG) – Sexual violence and geographies of children and adolescents in Paraná.
Samuel Silva Chaves (PhD, UEPG) – Spatial analysis of school proficiency and socioeconomic inequality in Brazilian capitals.
Déris Warmuth (PhD, UNICENTRO) – Social entrepreneurship and women's work in the Quilombo Paiol de Telha.
The software is used by two research groups actively working in graduate programs:

GETE – Grupo de Estudos Territoriais (Territorial Studies Group) at UEPG.
GEPES – Grupo de Pesquisa Redes de Poder, Migrações e Dinâmicas Territoriais (Research Group on Power Networks, Migrations, and Territorial Dynamics) at UNICENTRO.
SATA provides researchers with a GUI and tools for text filtering, bigram extraction, gender classification, and lexical/statistical analysis, including outputs ready for semantic network analysis in Gephi. If helpful, I would be glad to expand the documentation or provide additional usage examples.
