### 2025-07-01 344f75b
Fix tooltip bbox error and add newline
