### 2025-07-17 #8599 @editorialbot
[REVIEW]: ATAS - Academic Text Analysis System
**Submitting author:** @AlidesChimin (<a href="http://orcid.org/0000-0002-7436-390X">Alides Baptista Chimin Junior</a>)
**Repository:** https://github.com/AlidesChimin/SATA
**Branch with paper.md** (empty if default branch): main
**Version:** 1.0.0
