### 2025-08-13 #8599 @DiegoAscanio
Well, I believe that for the current status of Brazilian social sciences universe, this software is useful in helping social science researchers, mainly those who have some issues in developing / using computer programs to obtain insights and useful information like tag clouds, gender info, table data, etc from text documents they should study.

As a Brazilian STEM researcher who occasionally contributes with social researchers  at their projects, I confirm that many of the researchers complains about technical skills needed to perform their projects and this software might help them addressing their daily tasks, as its author claims.

It represents a meaningful contribution for Brazilian social science researchers and for that reason, I'd recommend this paper for publication in JOSS with the following adjustments:

1. References fixes should be performed in order to make the paper appropriate for publication. I ask the author to read the joss paper example hosted here so he can make the appropriate citations [https://joss.readthedocs.io/en/latest/example_paper.html](https://joss.readthedocs.io/en/latest/example_paper.html)

2. Statement of need: I particularly did not understand how GIS limitations are correlated to the problem the author tries to solve here, it didn't provided clarity to put this work into a wide context ahead of the literature. There are barriers and skill gaps which make it difficult to social science researchers adopt computational tools into their work routine, as described by [Metzler et al. (2016)](https://doi.org/10.4135/wp160926) and as I could understand, the author's software helps researchers in overcoming such barriers right? So I recommend this adaptation in the Statement of Need section in order to proper situate the paper's contributions.

Therefore, I recommend this paper for publication, provided that the minor revisions outlined above are addressed.
