### 2025-08-13 a49674b
Update paper.md
