### 2025-08-13 7aca766
Update paper.bib
