+++ paper.bib
@@ -1,118 +1,126 @@
 @book{Bardin:2011,
-author = {Bardin, Laurence},
-title = {Análise de Conteúdo},
-publisher = {Edições 70},
-year = {2011},
-address = {Lisboa}
+  author    = {Laurence Bardin},
+  title     = {Análise de Conteúdo},
+  publisher = {Edições 70},
+  address   = {Lisboa},
+  year      = {2011}
 }
 
-@article{Camara:2017,
-author = {Câmara, Gilberto},
-title = {Reflexões sobre a Geoinformação},
-journal = {Revista Brasileira de Cartografia},
-year = {2017},
-volume = {69},
-number = {5},
-pages = {821-830},
-doi = {10.14393/rbcv69n5-50836},
-url = {https://doi.org/10.14393/rbcv69n5-50836}
+@article{Camara:2017a,
+  author    = {Gilberto Câmara},
+  title     = {Reflexões sobre a Geoinformação},
+  journal   = {Revista Brasileira de Cartografia},
+  volume    = {69},
+  number    = {5},
+  pages     = {821--830},
+  year      = {2017},
+  doi       = {10.14393/rbcv69n5-50836},
+  url       = {https://doi.org/10.14393/rbcv69n5-50836}
 }
 
-@article{silva2017,
-author = {Edson Armando Silva and João Paulo Leandro de Almeida and Joseli Maria Silva and Mayã Polo de Campos},
-title = {Técnicas de Análise de Conteúdo: Experiências de Pesquisas Desenvolvidas pelo Grupo de Estudos Territoriais},
-journal = {Revista Latino Americana de Geografia e Gênero},
-volume = {8},
-number = {2},
-pages = {401-425},
-year = {2017},
-issn = {2177-2886}
+@misc{Camara:2017b,
+  author    = {Gilberto Câmara and Antônio Miguel Vieira Monteiro and João S. Medeiros},
+  title     = {Fundamentos Epistemológicos da Ciência da Geoinformação},
+  year      = {2017},
+  url       = {http://www.dpi.inpe.br/gilberto/livro/},
+  note      = {Accessed: 2017-10}
 }
 
-@book{barbetta2007,
-author = {Pedro A. Barbetta},
-title = {Estatística Aplicada às Ciências Sociais},
-edition = {7},
-publisher = {Ed. UFSC},
-year = {2007},
-address = {Florianópolis}
+@misc{Camara:2007,
+  author    = {Gilberto Câmara and Antônio Miguel Vieira Monteiro},
+  title     = {Conceitos Básicos em Ciência da Geoinformação},
+  year      = {2007},
+  url       = {http://www.dpi.inpe.br/gilberto/livro/},
+  note      = {Accessed: 2017-10}
 }
 
-@online{camara2007,
-author = {Gilberto Câmara and Antônio Miguel Vieira Monteiro},
-title = {Conceitos Básicos em Ciência da Geoinformação},
-year = {2007},
-url = {http://www.dpi.inpe.br/gilberto/livro/},
-note = {Acessado em outubro de 2017}
+@article{Silva:2017,
+  author    = {Edson Armando Silva and João Paulo Leandro de Almeida and Joseli Maria Silva and Mayã Polo de Campos},
+  title     = {Técnicas de Análise de Conteúdo: Experiências de Pesquisas Desenvolvidas pelo Grupo de Estudos Territoriais},
+  journal   = {Revista Latino Americana de Geografia e Gênero},
+  volume    = {8},
+  number    = {2},
+  pages     = {401--425},
+  year      = {2017},
+  issn      = {2177-2886}
 }
 
-@online{camara2017,
-author = {Gilberto Câmara and Antônio Miguel Vieira Monteiro and João S. Medeiros},
[…]
