### 2025-08-13 e83339b
Update paper.bib
