+++ paper.bib
@@ -124,3 +124,37 @@
   doi       = {10.4135/wp160926},
   url       = {https://doi.org/10.4135/wp160926}
 }
+
+@book{Harvey:2005,
+  author    = {David Harvey},
+  title     = {A Produção Capitalista do Espaço},
+  year      = {2005},
+  publisher = {Annablume},
+  address   = {São Paulo},
+  note      = {Título original: Spaces of Capital: Towards a Critical Geography, 2001}
+}
+
+@book{Harvey:1980,
+  author    = {David Harvey},
+  title     = {A Justiça Social e a Cidade},
+  year      = {1980},
+  publisher = {Hucitec},
+  address   = {São Paulo},
+  note      = {Título original: Social Justice and the City, 1973}
+}
+
+@book{Santos:1996,
+  author    = {Milton Santos},
+  title     = {A Natureza do Espaço: Técnica e Tempo, Razão e Emoção},
+  year      = {1996},
+  publisher = {Hucitec},
+  address   = {São Paulo}
+}
+
+@book{Santos:1978,
+  author    = {Milton Santos},
+  title     = {Por Uma Geografia Nova: Da Crítica da Geografia a uma Geografia Crítica},
+  year      = {1978},
+  publisher = {Hucitec},
+  address   = {São Paulo}
+}
