### 2025-08-13 b3b8963
Update paper.md
