+++ paper.md
@@ -31,14 +31,9 @@ We emphasize that the tools are in Portuguese, as they were developed by a Brazi
 # Statement of Need
 
 In the Brazilian Humanities and Social Sciences research context, many scholars still rely on manual or semi-manual methods to process and analyze large textual corpora. According to **Metzler et al. [@Metzler:2016]**, barriers such as limited programming skills, lack of access to adequate infrastructure, and scarce training opportunities hinder the adoption of computational approaches in these fields. As a result, essential tasks like identifying thematic patterns, generating bigram networks, classifying authors by gender, and calculating lexical statistics often become labor-intensive and error-prone.
-
 The **ATAS – Academic Text Analysis System** addresses these issues by providing an open-source, graphical-interface-based solution that automates these processes without requiring advanced technical expertise. By integrating Natural Language Processing (NLP) capabilities in Brazilian Portuguese, ATAS bridges the gap between sophisticated analytical methods and their practical usability for non-technical researchers. This directly supports more equitable access to computational tools in contexts where language and resource limitations frequently exclude researchers from digital scholarship.
-
 Beyond facilitating traditional content analysis, ATAS expands methodological possibilities for examining the relationship between discourse and spatiality. While **Geographic Information Systems (GIS)** are powerful tools for spatial analysis and thematic cartography, their architecture — based on discrete vector and raster data structures — tends to produce static “snapshots” of space. This structural limitation often prevents them from representing the inherently dynamic, processual, and socially constructed nature of geographic phenomena.
-
-As highlighted by **David Harvey [2005; 1980]** and **Milton Santos [1996; 1978]**, geography extends far beyond cartographic representation. Space is not merely a neutral container of events, but a social, political, and economic construct permeated by power relations, symbolic appropriation, and subjective experiences — dimensions that resist reduction to numerical variables in a GIS database.
-
-**Gilberto Câmara [@Camara:2017a]** critiques this limitation by noting that GIS computational structures often oversimplify spatial dynamics and overlook the dialectics of geographic phenomena. ATAS offers a methodological alternative by enabling the extraction and analysis of “discursive spatialities” — the ways in which space is constructed, contested, and redefined through language — using spatial statistics and semantic networks directly from textual data. In doing so, it complements rather than replaces cartography, offering researchers in the Humanities and Social Sciences a way to capture spatial meaning that is processual, contextual, and deeply embedded in discourse.
+As highlighted by **David Harvey [2005; 1980]** and **Milton Santos [1996; 1978]**, geography extends far beyond cartographic representation. Space is not merely a neutral container of events, but a social, political, and economic construct permeated by power relations, symbolic appropriation, and subjective experiences — dimensions that resist reduction to numerical variables in a GIS database. **Gilberto Câmara [@Camara:2017a]** critiques this limitation by noting that GIS computational structures often oversimplify spatial dynamics and overlook the dialectics of geographic phenomena. ATAS offers a methodological alternative by enabling the extraction and analysis of “discursive spatialities” — the ways in which space is constructed, contested, and redefined through language — using spatial statistics and semantic networks directly from textual data. In doing so, it complements rather than replaces cartography, offering researchers in the Humanities and Social Sciences a way to capture spatial meaning that is processual, contextual, and deeply embedded in discourse.
 
 
 # Features and Usage
