### 2025-08-13 #8599 @AlidesChimin
> Bem, acredito que, no atual cenário do universo das ciências sociais brasileiras, este software é útil para auxiliar pesquisadores de ciências sociais, principalmente aqueles que têm dificuldades em desenvolver/usar programas de computador para obter insights e informações úteis, como nuvens de tags, informações de gênero, dados de tabelas, etc., de documentos de texto que devem estudar.
>
> Como pesquisador brasileiro de STEM que ocasionalmente contribui com pesquisadores sociais em seus projetos, confirmo que muitos pesquisadores reclamam das habilidades técnicas necessárias para executar seus projetos e que este software pode ajudá-los a lidar com suas tarefas diárias, como afirma seu autor.
>
> Representa uma contribuição significativa para os pesquisadores brasileiros das ciências sociais e, por essa razão, recomendo a publicação deste artigo na JOSS com os seguintes ajustes:
>
> 1. Correções nas referências devem ser feitas para tornar o artigo adequado para publicação. Peço ao autor que leia o exemplo de artigo da Joss hospedado aqui para que ele possa fazer as citações apropriadas https://joss.readthedocs.io/en/latest/example_paper.html
> 2. Declaração de necessidade: Eu particularmente não entendi como as limitações do SIG se correlacionam com o problema que o autor tenta resolver aqui. Isso não forneceu clareza para situar este trabalho em um contexto amplo e à frente da literatura. Existem barreiras e lacunas de habilidades que dificultam a adoção de ferramentas computacionais por pesquisadores de ciências sociais em sua rotina de trabalho, conforme descrito por [Metzler et al. (2016)](https://doi.org/10.4135/wp160926) e, pelo que pude entender, o software do autor auxilia os pesquisadores a superar essas barreiras, certo? Portanto, recomendo esta adaptação na seção Declaração de Necessidade para situar adequadamente as contribuições do artigo.
>
> Portanto, recomendo a publicação deste artigo, desde que as pequenas revisões descritas acima sejam abordadas.

Prezado(a) avaliador(a),

[paper.md](https://github.com/user-attachments/files/21759924/paper.md)

Agradeço imensamente pelas valiosas contribuições e observações feitas em seu parecer. Suas considerações foram fundamentais para aprimorar a clareza, a consistência e a fundamentação do manuscrito.
Atendendo às sugestões apresentadas, realizei as modificações demandadas, incluindo:
- Revisão e padronização completa das referências, conforme as diretrizes do JOSS, eliminando inconsistências e acrescentando as citações necessárias.
- Reestruturação e aprofundamento da seção Statement of Need, situando de forma mais clara o problema central que o software busca resolver, com ênfase nas barreiras identificadas por Metzler et al. (2016) e na fundamentação teórica que sustenta a crítica às limitações dos SIGs, conforme discutido por David Harvey, Milton Santos e Gilberto Câmara.

Estou confiante de que as alterações realizadas atendem plenamente às observações e fortalecem a contribuição do trabalho.
Agradeço novamente pela leitura atenta e pelas orientações, que foram determinantes para o aprimoramento do artigo.

Atenciosamente,
Alides Baptista Chimin Junior
