### 2025-08-17 #8599 @rafaelanchieta
In summary, the paper presents a software to extract some information from a text, such as part-of-speech tags, named entities, sentiment orientation, and so on. For that, the software uses mainly the spaCy and TextBlob tools.

Although the software may be helpful for Brazilian social science researchers, I suggest developing more flexible software. For example, instead of using only spaCy, the software could allow the user to select which tool to use for extracting information. There are several tools with similar functionalities to spaCy. For instance, NLPNet, Stanza, Udpipe, etc, were trained for the Brazilian Portuguese language.

I also suggest removing the TextBlob library. This tool was developed for the English language, and there is no support for Brazilian Portuguese [TextBlob](https://github.com/sloria/TextBlob/issues/202#issuecomment-388559818). There are several sentiment analysis models trained for Portuguese [Portuguese-NLP](https://github.com/ajdavidl/Portuguese-NLP).
