### 2025-08-18 dcc37c1
Update paper.md
