+++ paper.md
@@ -18,43 +18,39 @@ date: 18 February 2025
 bibliography: paper.bib
 ---
 
-
 # Abstract
 
-The **ATAS - Academic Text Analysis System** (Brazilian Portuguese SATA - Sistema de Análise de Textos Acadêmicos) is an open-source software developed to assist researchers in **textual content analysis**. Inspired by the methodology of **Laurence Bardin [@Bardin:2011]**, ATAS facilitates the **extraction, filtering, and statistical analysis** of academic texts, enabling the identification of semantic and linguistic patterns.
-
-The software is particularly useful for researchers in the **Humanities and Social Sciences**, providing tools for analyzing **keywords, bigrams, lexical categories**, and other quantitative metrics of textual analysis. It integrates natural language processing (NLP) and allows data export to software such as **Gephi** for semantic network analysis.
-
+The ATAS - Academic Text Analysis System (Brazilian Portuguese SATA - Sistema de Análise de Textos Acadêmicos) is an open-source software developed to assist researchers in textual content analysis. Inspired by the methodology of @Bardin:2011, ATAS facilitates the extraction, filtering, and statistical analysis of academic texts, enabling the identification of semantic and linguistic patterns.
+The software is particularly useful for researchers in the Humanities and Social Sciences, providing tools for analyzing keywords, bigrams, lexical categories, and other quantitative metrics of textual analysis. It integrates natural language processing (NLP) and allows data export to software such as @Gephi:2009 for semantic network analysis.
 We emphasize that the tools are in Portuguese, as they were developed by a Brazilian researcher and are being used by the GEPES and GETE research groups.
-
 
 # Statement of Need
 
-In the Brazilian Humanities and Social Sciences research context, many scholars still rely on manual or semi-manual methods to process and analyze large textual corpora. According to **Metzler et al. [@Metzler:2016]**, barriers such as limited programming skills, lack of access to adequate infrastructure, and scarce training opportunities hinder the adoption of computational approaches in these fields. As a result, essential tasks like identifying thematic patterns, generating bigram networks, classifying authors by gender, and calculating lexical statistics often become labor-intensive and error-prone.
-The **ATAS – Academic Text Analysis System** addresses these issues by providing an open-source, graphical-interface-based solution that automates these processes without requiring advanced technical expertise. By integrating Natural Language Processing (NLP) capabilities in Brazilian Portuguese, ATAS bridges the gap between sophisticated analytical methods and their practical usability for non-technical researchers. This directly supports more equitable access to computational tools in contexts where language and resource limitations frequently exclude researchers from digital scholarship.
-Beyond facilitating traditional content analysis, ATAS expands methodological possibilities for examining the relationship between discourse and spatiality. While **Geographic Information Systems (GIS)** are powerful tools for spatial analysis and thematic cartography, their architecture — based on discrete vector and raster data structures — tends to produce static “snapshots” of space. This structural limitation often prevents them from representing the inherently dynamic, processual, and socially constructed nature of geographic phenomena.
[…]
