### 2025-08-18 40d97ec
Update paper.bib
