+++ paper.bib
@@ -6,34 +6,6 @@
   year      = {2011}
 }
 
-@article{Camara:2017a,
-  author    = {Gilberto Câmara},
-  title     = {Reflexões sobre a Geoinformação},
-  journal   = {Revista Brasileira de Cartografia},
-  volume    = {69},
-  number    = {5},
-  pages     = {821--830},
-  year      = {2017},
-  doi       = {10.14393/rbcv69n5-50836},
-  url       = {https://doi.org/10.14393/rbcv69n5-50836}
-}
-
-@misc{Camara:2017b,
-  author    = {Gilberto Câmara and Antônio Miguel Vieira Monteiro and João S. Medeiros},
-  title     = {Fundamentos Epistemológicos da Ciência da Geoinformação},
-  year      = {2017},
-  url       = {http://www.dpi.inpe.br/gilberto/livro/},
-  note      = {Accessed: 2017-10}
-}
-
-@misc{Camara:2007,
-  author    = {Gilberto Câmara and Antônio Miguel Vieira Monteiro},
-  title     = {Conceitos Básicos em Ciência da Geoinformação},
-  year      = {2007},
-  url       = {http://www.dpi.inpe.br/gilberto/livro/},
-  note      = {Accessed: 2017-10}
-}
-
 @article{Silva:2017,
   author    = {Edson Armando Silva and João Paulo Leandro de Almeida and Joseli Maria Silva and Mayã Polo de Campos},
   title     = {Técnicas de Análise de Conteúdo: Experiências de Pesquisas Desenvolvidas pelo Grupo de Estudos Territoriais},
@@ -63,14 +35,6 @@
   pages     = {201--210},
   year      = {2006},
   address   = {Brasília}
-}
-
-@software{TextBlob:2018,
-  author    = {Steven Loria},
-  title     = {TextBlob: Simplified Text Processing},
-  year      = {2018},
-  url       = {https://textblob.readthedocs.io/en/dev/},
-  note      = {Accessed: 2025-02}
 }
 
 @software{spaCy:2023,
