### 2025-09-07 #8599 @felipemaiapolo
Hi @RichardLitt ,

Thank you for the reminder!

The proposed tool has the potential to have an impact on the Brazilian social sciences research landscape. However, I would recommend that the author contextualize their work in more detail and with more recent references. The justification of the work cites related work (e.g., @Metzler:2016) that was published years ago. As we all know, the field of NLP and AI has undergone a huge shift in the last 3 years with the advent of LLMs. As far as I know, this has also impacted many fields and how they analyze text. However, none of this is present in the related work section. It would be interesting to see how the author sees this big shift and how their work can benefit from the recent advances in the text analysis field. I would also recommend the addition of some of these new tools in SATA if possible (tools such as https://openrouter.ai/ allow free use of LLMs and could be used to make the work more impactful, for example).

As a result, I do not recommend this work for acceptance before some consideration of the more recent context.

I am open for discussions in case the authors want to talk more about it.
