### 2025-10-29 17bef01
Add CONTRIBUTING.md for project guidelines
