+++ CONTRIBUTING.md
@@ -1,0 +1,26 @@
+# Contributing to ATAS - Academic Text Analysis System
+
+Thank you for your interest in contributing to the ATAS project!
+
+## How to Contribute
+
+1. Fork the repository and create your branch from `main`.
+2. Make your changes and ensure the application runs locally.
+3. Commit your changes with clear messages.
+4. Open a Pull Request with a detailed description.
+
+## Code Style
+
+Please follow standard PEP8 style for Python code.
+
+## Reporting Issues
+
+If you find a bug or have a suggestion, please open an issue using the GitHub issue tracker.
+
+## Feature Suggestions
+
+Feature requests are welcome! Feel free to discuss them through issues before starting implementation.
+
+---
+
+ATAS is an academic open-source project maintained by Alides Chimin. Let's work together to make it even better!
