### 2025-10-29 e96fe15
Revise CONTRIBUTING.md for improved instructions
Updated the contributing guidelines for clarity and conciseness.
