### 2025-10-29 314a036
Revise CONTRIBUTING.md with new contribution guidelines
Updated the contributing guidelines to include new sections on ways to contribute and getting started with local development.
