### 2025-10-29 5d8722f
Revise CONTRIBUTING.md for clarity and structure
