### 2025-10-29 86d3851
Update README to include contribution guidelines
Added a note to refer to CONTRIBUTING.md for contributions.
