+++ README.md
@@ -18,3 +18,5 @@ pip3 install -r requirements.txt
 python3 src/main.py
 ```
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.14868064.svg)](https://zenodo.org/records/14868064)
+
+Veja CONTRIBUTING.md para contribuir
