### 2025-10-29 dabd1c7
Add GitHub Actions workflow for testing
