+++ .github/workflows/tests.yml
@@ -1,0 +1,32 @@
+name: Tests
+
+on:
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ main ]
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.11'
+
+      - name: Install deps
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r requirements.txt
+          pip install pytest
+
+      # If a spaCy model is required by tests, uncomment and adjust:
+      # - name: Install spaCy model (optional)
+      #   run: pip install src/pt_core_news_sm-3.8.0-py3-none-any.whl
+
+      - name: Run tests
+        run: pytest -q
