### 2025-10-29 990de0e
Add import tests for processamento package
Added minimal import tests for the processamento package and its modules.
