### 2025-10-29 c6ad544
Add test status badge to README
Added a badge for test status to the README.
