### 2025-10-29 #8599 @AlidesChimin
> [@AlidesChimin](https://github.com/AlidesChimin)Não estou encontrando nenhuma documentação de contribuição, nem testes automatizados. Estou perdendo alguma coisa?

Added CONTRIBUTING.md, a pytest suite and GitHub Actions CI (.github/workflows/tests.yml). Tests pass on main. We’re ready for the next review step. We can expand tests further if needed.
