### 2025-11-05 ac5f87f
Fix formatting and add new references in paper.bib
Updated titles and authors in various entries for correct accentuation and formatting. Added new references for recent publications.
