### 2025-11-05 a6507e2
Fix formatting and update entries in paper.bib
