+++ paper.bib
@@ -1,23 +1,15 @@
 @book{Bardin:2011,
   author    = {Laurence Bardin},
-  title     = {An{\'a}lise de Conte{\'u}do},
-  publisher = {Edi{\c c}{\~o}es 70},
+  title     = {Análise de Conteúdo},
+  publisher = {Edições 70},
   address   = {Lisboa},
   year      = {2011}
 }
 
-@book{Bardin:2020,
-  author    = {Laurence Bardin},
-  title     = {An{\'a}lise de Conte{\'u}do},
-  publisher = {Edi{\c c}{\~o}es 70},
-  address   = {S{\~a}o Paulo},
-  year      = {2020}
-}
-
 @article{Silva:2017,
-  author    = {Edson Armando Silva and Jo{\~a}o Paulo Leandro de Almeida and Joseli Maria Silva and May{\~a} Polo de Campos},
-  title     = {T{\'e}cnicas de An{\'a}lise de Conte{\'u}do: Experi{\^e}ncias de Pesquisas Desenvolvidas pelo Grupo de Estudos Territoriais},
-  journal   = {Revista Latino Americana de Geografia e G{\^e}nero},
+  author    = {Edson Armando Silva and João Paulo Leandro de Almeida and Joseli Maria Silva and Mayã Polo de Campos},
+  title     = {Técnicas de Análise de Conteúdo: Experiências de Pesquisas Desenvolvidas pelo Grupo de Estudos Territoriais},
+  journal   = {Revista Latino Americana de Geografia e Gênero},
   volume    = {8},
   number    = {2},
   pages     = {401--425},
@@ -27,62 +19,64 @@
 
 @book{Barbetta:2007,
   author    = {Pedro A. Barbetta},
-  title     = {Estat{\'\i}stica Aplicada {\`a}s Ci{\^e}ncias Sociais},
+  title     = {Estatística Aplicada às Ciências Sociais},
   edition   = {7},
   publisher = {Ed. UFSC},
-  address   = {Florian{\'o}polis},
+  address   = {Florianópolis},
   year      = {2007}
 }
 
 @article{Gunther:2006,
-  author    = {Hartmut G{\"u}nther},
-  title     = {Pesquisa Qualitativa Versus Pesquisa Quantitativa: Esta {\'e} a Quest{\~a}o?},
+  author    = {Hartmut Günther},
+  title     = {Pesquisa Qualitativa Versus Pesquisa Quantitativa: Esta é a Questão?},
   journal   = {Revista Psicologia: Teoria e Pesquisa},
   volume    = {22},
   number    = {2},
   pages     = {201--210},
   year      = {2006},
-  address   = {Bras{\'\i}lia}
+  address   = {Brasília}
 }
 
-@misc{spaCy:2023,
-  author       = {{Explosion AI}},
-  title        = {spaCy: Industrial-Strength Natural Language Processing in {Python}},
-  year         = {2023},
-  howpublished = {\url{https://spacy.io/}},
-  note         = {Version 3.8. Accessed: 2025-02}
+@software{spaCy:2023,
+  author    = {Explosion AI},
+  title     = {spaCy: Industrial-Strength Natural Language Processing in Python},
+  year      = {2023},
+  version   = {3.7.0},
+  url       = {https://spacy.io/},
+  note      = {Accessed: 2025-02}
 }
 
-@misc{Pandas:2023,
-  author       = {{The Pandas Development Team}},
-  title        = {pandas: {Python} Data Analysis Library},
-  year         = {2023},
-  howpublished = {\url{https://pandas.pydata.org/}},
-  note         = {Version 2.3. Accessed: 2025-02}
+@software{Pandas:2023,
+  author    = {{The Pandas Development Team}},
+  title     = {pandas: Python Data Analysis Library},
[…]
