### 2025-11-05 0246240
Update paper.md
