### 2025-11-05 d2606d9
Revise bibliography with new and updated references
Updated bibliographic entries with new publications and corrected details.
