+++ paper.bib
@@ -1,82 +1,9 @@
-@book{Bardin:2011,
+@book{Bardin:2020,
   author    = {Laurence Bardin},
-  title     = {Análise de Conteúdo},
-  publisher = {Edições 70},
-  address   = {Lisboa},
-  year      = {2011}
-}
-
-@article{Silva:2017,
-  author    = {Edson Armando Silva and João Paulo Leandro de Almeida and Joseli Maria Silva and Mayã Polo de Campos},
-  title     = {Técnicas de Análise de Conteúdo: Experiências de Pesquisas Desenvolvidas pelo Grupo de Estudos Territoriais},
-  journal   = {Revista Latino Americana de Geografia e Gênero},
-  volume    = {8},
-  number    = {2},
-  pages     = {401--425},
-  year      = {2017},
-  issn      = {2177-2886}
-}
-
-@book{Barbetta:2007,
-  author    = {Pedro A. Barbetta},
-  title     = {Estatística Aplicada às Ciências Sociais},
-  edition   = {7},
-  publisher = {Ed. UFSC},
-  address   = {Florianópolis},
-  year      = {2007}
-}
-
-@article{Gunther:2006,
-  author    = {Hartmut Günther},
-  title     = {Pesquisa Qualitativa Versus Pesquisa Quantitativa: Esta é a Questão?},
-  journal   = {Revista Psicologia: Teoria e Pesquisa},
-  volume    = {22},
-  number    = {2},
-  pages     = {201--210},
-  year      = {2006},
-  address   = {Brasília}
-}
-
-@software{spaCy:2023,
-  author    = {Explosion AI},
-  title     = {spaCy: Industrial-Strength Natural Language Processing in Python},
-  year      = {2023},
-  version   = {3.7.0},
-  url       = {https://spacy.io/},
-  note      = {Accessed: 2025-02}
-}
-
-@software{Pandas:2023,
-  author    = {{The Pandas Development Team}},
-  title     = {pandas: Python Data Analysis Library},
-  year      = {2023},
-  version   = {2.2.0},
-  url       = {https://pandas.pydata.org/},
-  note      = {Accessed: 2025-02}
-}
-
-@software{GenderGuesser:2016,
-  author    = {Daniel Hutchinson},
-  title     = {gender-guesser: Library for Guessing the Gender of First Names},
-  year      = {2016},
-  url       = {https://pypi.org/project/gender-guesser/},
-  note      = {Accessed: 2025-02}
-}
-
-@software{Gephi:2009,
-  author    = {Mathieu Bastian and Sebastien Heymann and Mathieu Jacomy},
-  title     = {Gephi: An Open Source Software for Exploring and Manipulating Networks},
-  year      = {2009},
-  url       = {https://gephi.org/},
-  note      = {Accessed: 2025-02}
-}
-
-@software{Tkinter:2023,
-  author    = {Guido van Rossum and {The Python Software Foundation}},
-  title     = {Tkinter: Python Interface to Tcl/Tk},
-  year      = {2023},
-  url       = {https://docs.python.org/3/library/tkinter.html},
-  note      = {Accessed: 2025-02}
+  title     = {An{\'a}lise de Conte{\'u}do},
+  publisher = {Edi{\c c}{\~o}es 70},
+  address   = {S{\~a}o Paulo},
+  year      = {2020}
 }
 
 @misc{Metzler:2016,
@@ -89,36 +16,93 @@
   url       = {https://doi.org/10.4135/wp160926}
 }
 
-@book{Harvey:2005,
-  author    = {David Harvey},
-  title     = {A Produção Capitalista do Espaço},
+@book{Massey:2005,
+  author    = {Doreen Massey},
+  title     = {For Space},
[…]
