### 2025-11-05 #8599 @AlidesChimin
> Hi [@RichardLitt](https://github.com/RichardLitt) ,
>
> Thank you for the reminder!
>
> The proposed tool has the potential to have an impact on the Brazilian social sciences research landscape. However, I would recommend that the author contextualize their work in more detail and with more recent references. The justification of the work cites related work (e.g., [@Metzler](https://github.com/Metzler):2016) that was published years ago. As we all know, the field of NLP and AI has undergone a huge shift in the last 3 years with the advent of LLMs. As far as I know, this has also impacted many fields and how they analyze text. However, none of this is present in the related work section. It would be interesting to see how the author sees this big shift and how their work can benefit from the recent advances in the text analysis field. I would also recommend the addition of some of these new tools in SATA if possible (tools such as https://openrouter.ai/ allow free use of LLMs and could be used to make the work more impactful, for example).
>
> As a result, I do not recommend this work for acceptance before some consideration of the more recent context.
>
> I am open for discussions in case the authors want to talk more about it.

Dear reviewer [@RichardLitt],

Thank you very much for your careful reading and for highlighting this important point about contextualizing the work within the current NLP and LLM landscape.

Following your suggestion, we have substantially revised the “Statement of Need” to include recent developments in Large Language Models (LLMs) and the ways in which these have reshaped the field of computational text analysis. The updated version now cites and discusses recent open-access contributions such as BLOOM [@Scao:2022], LLaMA 2 [@Touvron:2023], and GPT-4 [@OpenAI:2023], while also situating the ATAS system in relation to the emerging ecosystem of Brazilian Portuguese NLP models, including BERTimbau [@Souza:2020] and PT-BR Encoders [@Mello:2024].

The revised text clarifies that ATAS is designed to complement, not compete with LLMs — offering a transparent, open-source, and fully local tool that promotes reproducibility and equitable access, particularly for non-English-speaking researchers in the Humanities and Social Sciences.

Moreover, we added a note in the “Features and Usage” section explaining that future versions of ATAS will integrate open LLMs fine-tuned for Portuguese, following the principles of openness and independence from proprietary APIs (e.g., OpenRouter).

We appreciate your valuable feedback, which helped us strengthen both the theoretical and technological framing of the work. The revised version now provides a clearer connection between ATAS and the recent advances in NLP, ensuring its continued relevance for digital scholarship in Brazil and beyond.

Thank you again for your thoughtful review and for contributing to the improvement of this submission.

Best regards,
Alides Baptista Chimin Junior
Universidade Estadual do Centro-Oeste (UNICENTRO)
