### 2025-11-05 25c6bf1
Revise paper content for clarity and structure
Updated text for clarity and consistency, improved phrasing, and added new sections on graphical interface, installation, reproducibility, and community guidelines.
