+++ paper.md
@@ -26,11 +26,11 @@ The toolset is localized for **Brazilian Portuguese** and is actively used by th
 
 # Statement of Need
 
-In Brazil’s Humanities and Social Sciences, many scholars still rely on manual or semi-manual workflows to process large textual corpora. Historical barriers—limited programming experience, scarce infrastructure, and uneven training—have hindered the adoption of computational methods [@Metzler:2016]. At the same time, **Large Language Models (LLMs)** have transformed NLP in recent years, opening powerful possibilities but also raising concerns about **language coverage, transparency, and reproducibility** [@OpenAI:2023; @Scao:2022; @Touvron:2023]. For **Portuguese (Brazil)** specifically, the ecosystem has improved with **encoder models** such as **BERTimbau** [@Souza:2020] and more recent PT-BR encoders [@Mello:2024], yet accessible, GUI-driven tools that operationalize these advances for **non-specialists** remain scarce.
+In Brazil’s Humanities and Social Sciences (Geography), many scholars still rely on manual or semi-manual workflows to process large textual corpora. Historical barriers—limited programming experience, scarce infrastructure, and uneven training—have hindered the adoption of computational methods [@Metzler:2016]. At the same time, **Large Language Models (LLMs)** have transformed NLP in recent years, opening powerful possibilities but also raising concerns about **language coverage, transparency, and reproducibility** [@Scao:2022; @OpenAI:2023]. For **Portuguese (Brazil)** specifically, the ecosystem has improved with **encoder models** such as **BERTimbau** [@Souza:2020], yet accessible, GUI-driven tools that operationalize these advances for **non-specialists** remain scarce.
 
-**ATAS** addresses this gap as an **open-source, GUI-based** system that automates key tasks—keyword filtering, bigram networks, lexical metrics—**without requiring advanced coding**. Localization to **Brazilian Portuguese** promotes equitable access to computational analysis and supports reproducible research workflows in Human Geography and related fields.
+**ATAS** addresses this gap as an **open-source, GUI-based** system that automates key tasks—keyword filtering, bigram networks, lexical metrics—**without requiring advanced coding**. Localization to **Brazilian Portuguese** promotes equitable access to computational analysis and supports reproducible research pipelines for Human Geography and allied fields.
 
-Beyond facilitating content analysis, ATAS helps investigate how **discourse constitutes space**. In Human Geography, **space** is socially produced and relational rather than a neutral container [@Lefebvre:1991; @Massey:2005]. While GIS excels at mapping and spatial querying, its discrete data structures can under-represent **processual and discursive spatialities**. ATAS complements cartography by extracting **semantic networks and entities** directly from text, enabling researchers to track how places, actors, and relations are constructed, contested, and reconfigured in academic and policy discourse.
+Beyond facilitating content analysis, ATAS helps investigate how **discourse constitutes space**. In Human Geography, **space** is socially produced and relational rather than a neutral container [@Harvey:1980; @Massey:2005]. While GIS excels at mapping and spatial querying, its discrete data structures can under-represent **processual and discursive spatialities**. ATAS complements cartography by extracting **semantic networks and entities** directly from text, enabling researchers to track how places, actors, and relations are constructed, contested, and reconfigured in academic and policy discourse.
 
 # Features and Usage
 
[…]
