### 2025-11-05 77dec1a
Revise scholarly effort section with updated metrics
Updated project description and statistics in paper.md.
