+++ examples/README.md
@@ -1,1 +1,39 @@
-teste
+# Examples for ATAS – Academic Text Analysis System
+
+This folder contains real data used to test ATAS functionalities.
+The corpus comes from the official transcript of the **Brazilian Chamber of Deputies**
+during the **2016 Impeachment of President Dilma Rousseff**.
+
+## 1. Input Texts
+
+- `Texto_Limpo.txt`: cleaned transcription of the parliamentary session.
+- `Video_Impeatchment_Dilma.txt`: alternative version extracted from video captions.
+
+## 2. Output Data (Generated by ATAS)
+
+| File | Description |
+|------|--------------|
+| `tabela.csv` | Bigrams extracted from text. |
+| `tabela_locais.csv` | Named entities – locations. |
+| `tabela_nomes.csv` | Named entities – people. |
+| `tabela_organizacao.csv` | Named entities – organizations. |
+| `estatistica.csv` | Text statistics: frequency, lexical diversity, entity counts. |
+| `estatistica_locais.csv` | Statistics by location. |
+| `estatistica_sentimento.csv` | Sentiment classification summary. |
+
+## 3. Reproduction Instructions
+
+Each example can be reproduced with:
+
+```bash
+# 1. Run text filtering
+python src/filtro_texto.py examples/Texto_Limpo.txt
+
+# 2. Convert to bigram table
+python src/converter_tabela.py examples/Texto_Limpo.txt
+
+# 3. Generate text statistics
+python src/estatisticas_texto.py examples/Texto_Limpo.txt
+
+# 4. Identify gender (if applicable)
+python src/identificador_sexo.py examples/tabela_nomes.csv
