+++ tests/test_import.py
@@ -1,0 +1,6 @@
+def test_import():
+    import filtro_texto
+    import converter_tabela
+    import estatisticas_texto
+    import identificador_sexo
+    assert True
