### 2025-11-05 c9d4f6e
Revise README for clarity and structure
Updated README to clarify data sources and outputs of ATAS. Enhanced structure and added sections for input texts, output tables, semantic network, word clouds, and reproduction instructions.
