+++ examples/README.md
@@ -1,39 +1,78 @@
 # Examples for ATAS – Academic Text Analysis System
 
-This folder contains real data used to test ATAS functionalities.
-The corpus comes from the official transcript of the **Brazilian Chamber of Deputies**
-during the **2016 Impeachment of President Dilma Rousseff**.
+This folder contains **real data and outputs** generated by the ATAS software (Academic Text Analysis System).  
+The dataset used comes from the **official transcript of the Brazilian Chamber of Deputies** during the **2016 impeachment session of President Dilma Rousseff**.
 
-## 1. Input Texts
+The files here demonstrate the **full workflow** of ATAS: text preprocessing, lexical filtering, statistical analysis, sentiment evaluation, and network generation (Gephi).
 
-- `Texto_Limpo.txt`: cleaned transcription of the parliamentary session.
-- `Video_Impeatchment_Dilma.txt`: alternative version extracted from video captions.
+---
 
-## 2. Output Data (Generated by ATAS)
+## 🧾 1. Input Texts
 
 | File | Description |
 |------|--------------|
-| `tabela.csv` | Bigrams extracted from text. |
-| `tabela_locais.csv` | Named entities – locations. |
-| `tabela_nomes.csv` | Named entities – people. |
-| `tabela_organizacao.csv` | Named entities – organizations. |
-| `estatistica.csv` | Text statistics: frequency, lexical diversity, entity counts. |
-| `estatistica_locais.csv` | Statistics by location. |
-| `estatistica_sentimento.csv` | Sentiment classification summary. |
+| `Texto_Limpo.txt` | Cleaned transcript of the 2016 impeachment session (source text). |
+| `Video_Impeatchment_Dilma.txt` | Alternative transcript extracted from video captions of the same event. |
 
-## 3. Reproduction Instructions
+---
 
-Each example can be reproduced with:
+## 📊 2. Output Tables (Generated by ATAS)
+
+| File | Description |
+|------|--------------|
+| `tabela.csv` | Bigrams extracted from the full text. |
+| `tabela_locais.csv` | Named entities recognized as **locations**. |
+| `tabela_nomes.csv` | Named entities recognized as **people**. |
+| `tabela_organizacao.csv` | Named entities recognized as **organizations**. |
+| `estatistica.csv` | Global lexical statistics: token frequency, lexical diversity, and word counts. |
+| `estatistica_locais.csv` | Statistical summary of geographical entities. |
+| `estatistica_sentimento.csv` | Preliminary sentiment analysis results for each speaker. |
+| `Nos.csv` | Node table exported from Gephi (5178 nodes). |
+| `Arestas.csv` | Edge table exported from Gephi (21961 edges). |
+
+---
+
+## 🌐 3. Semantic Network
+
+| File | Description |
+|------|--------------|
+| `Rede.gephi` | Network file containing 5178 nodes and 21961 edges, generated by ATAS and visualized in Gephi. |
+
+This network maps relationships among **key terms, people, institutions, and spatial references** mentioned during the parliamentary debate.
+
+---
+
+## 🖼️ 4. Word Clouds (Generated by ATAS)
+
+| Image | Description |
+|--------|-------------|
+| `estatistica_nuvem_adjetivos.png` | Word cloud of adjectives. |
+| `estatistica_nuvem_adverbios.png` | Word cloud of adverbs. |
+| `estatistica_nuvem_locais.png` | Word cloud of locations. |
+| `estatistica_nuvem_organizacoes.png` | Word cloud of organizations. |
+| `estatistica_nuvem_substantivos.png` | Word cloud of nouns. |
[…]
