### 2025-11-05 #8599 @AlidesChimin
Hi @RichardLitt,

All reviewer recommendations have been implemented. The Statement of Need was rewritten with recent references on NLP and LLMs, expanding the Human Geography framing and methodological discussion. The repository now includes a /examples folder with reproducible datasets (CSV tables, word clouds, and a Rede.gephi file with 5,178 nodes and 21,961 edges) generated from real text analysis.

A continuous integration workflow was added with automated tests (pytest) for all modules.
The latest run reports:

```
2 passed, 0 failed (3.48s)
```

Verification badge:
![Tests](https://github.com/AlidesChimin/SATA/actions/workflows/tests.yml/badge.svg)

Documentation, contributing, and conduct guidelines were also updated for transparency and reproducibility.
I remain available for any further adjustments or clarifications.

Best,
Alides Baptista Chimin Junior
