### 2025-11-05 #8599 @RichardLitt
@DiegoAscanio, @rafaelanchieta, and @felipemaiapolo: Substantial changes have been made. Can you rereview this paper?

@AlidesChimin I am seeing one broken reference in the rendered paper, and note the missing DOI above. Can you fix that? Thank you.
