### 2025-11-06 287351f
Update paper.bib
