### 2025-11-06 63c410e
Revise references in ATAS discourse analysis section
Updated references in the discussion of ATAS and its role in Human Geography.
