### 2025-11-06 be9b932
Update paper.bib
