+++ paper.bib
@@ -37,6 +37,7 @@
   author    = {Mathieu Bastian and Sebastien Heymann and Mathieu Jacomy},
   title     = {Gephi: An Open Source Software for Exploring and Manipulating Networks},
   year      = {2009},
+  doi       = {10.1609/icwsm.v3i1.13937},
   url       = {https://doi.org/10.1609/icwsm.v3i1.13937}
 }
 
