### 2025-11-06 7c709dc
Update paper.md
