### 2025-11-12 #8599 @RichardLitt
@DiegoAscanio, @rafaelanchieta, and @felipemaiapolo: Gentle ping. Please let me know if you can take another look in the next week.
