### 2025-11-18 #8599 @RichardLitt
Muito obrigado, Rafael.
