### 2025-11-19 #8599 @DiegoAscanio
As a Brazilian researcher with training in multiple fields (computer science, mathematical modeling, education, pedagogy, DEI, etc.), I can attest that many researchers in the social sciences come from contexts marked by deep inequality, poverty, and lack of opportunity. These conditions often create additional barriers to learning new technologies and concepts related to computer programming, mathematics, statistics, and related areas.

These researchers can and do overcome such obstacles; however, with tools properly tailored to their needs, they can do so more efficiently, while achieving greater progress in their research goals. In this sense, the relevance of ATAS for Brazilian social sciences is clear. Moreover, its importance is not restricted to the Brazilian context, but extends to other Lusophone countries in development, which often face even more severe challenges due to underdevelopment and underinvestment in research and education.

As the author has addressed all issues raised by the reviewers, and as both the software and the paper satisfy the JOSS publication criteria, I therefore reaffirm my recommendation that this paper be accepted for publication.
