### 2025-11-23 #8599 @felipemaiapolo
I would like to acknowledge the authors' effort in mentioning LLMs in their related work section. I am in favor of the paper acceptance. Thank you.
