### 2025-12-02 #8599 @RichardLitt
@AlidesChimin Before minting the new release, can you remove all of the bolded keywords in the text? They make reading rather difficult for me, and I think they are unnecessary.
