### 2025-12-02 cd173c5
docs: Remove unnecessary line edits
