### 2025-12-02 #8599 @RichardLitt
Small related PR in https://github.com/AlidesChimin/SATA/pull/7/files.
