### 2025-12-03 cdf3f17
Remove HTML entities and clean up text formatting
