+++ paper.md
@@ -20,17 +20,17 @@ bibliography: paper.bib
 
 # Abstract
 
-The **ATAS – Academic Text Analysis System** (Brazilian Portuguese **SATA – Sistema de Análise de Textos Acadêmicos**) is open-source software to support **textual content analysis**. Inspired by **Bardin’s method** [@Bardin:2020], ATAS streamlines **extraction, filtering, and statistical analysis** of academic texts to surface semantic and linguistic patterns. It is particularly useful to **Humanities and Social Sciences (Geography)** researchers, offering tools for **keywords, bigrams, lexical categories**, and other quantitative text measures. ATAS integrates **NLP** workflows and exports data to tools such as **Gephi** for semantic network analysis [@Gephi:2009].
+The ATAS – Academic Text Analysis System (Brazilian Portuguese SATA – Sistema de Análise de Textos Acadêmicos) is open-source software to support textual content analysis. Inspired by Bardin’s method [@Bardin:2020], ATAS streamlines extraction, filtering, and statistical analysis of academic texts to surface semantic and linguistic patterns. It is particularly useful to Humanities and Social Sciences (Geography) researchers, offering tools for keywords, bigrams, lexical categories, and other quantitative text measures. ATAS integrates NLP workflows and exports data to tools such as Gephi for semantic network analysis [@Gephi:2009].
 
-The toolset is localized for **Brazilian Portuguese** and is actively used by the **GETE – Grupo de Estudos Territoriais (Territorial Studies Group)** at **UEPG** and the **GEPES – Grupo de Pesquisa Redes de Poder, Migrações e Dinâmicas Territoriais (Research Group on Power Networks, Migrations, and Territorial Dynamics)** at **UNICENTRO** (Brazil), both operating in **Master’s and PhD** programs.
+The toolset is localized for Brazilian Portuguese and is actively used by the GETE – Grupo de Estudos Territoriais (Territorial Studies Group) at UEPG and the GEPES – Grupo de Pesquisa Redes de Poder, Migrações e Dinâmicas Territoriais (Research Group on Power Networks, Migrations, and Territorial Dynamics) at UNICENTRO (Brazil), both operating in Master’s and PhD programs.
 
 # Statement of Need
 
-In Brazil’s Humanities and Social Sciences (Geography), many scholars still rely on manual or semi-manual workflows to process large textual corpora. Historical barriers—limited programming experience, scarce infrastructure, and uneven training—have hindered the adoption of computational methods [@Metzler:2016]. At the same time, **Large Language Models (LLMs)** have transformed NLP in recent years, opening powerful possibilities but also raising concerns about **language coverage, transparency, and reproducibility** [@Scao:2022; @OpenAI:2023]. For **Portuguese (Brazil)** specifically, the ecosystem has improved with **encoder models** such as **BERTimbau** [@Souza:2020] and more recent **PT-BR Encoders** [@Mello:2024], yet accessible, GUI-driven tools that operationalize these advances for **non-specialists** remain scarce.
[…]
