### 2025-12-03 7e3322c
Add Figshare DOI to paper.md
