### 2025-12-04 #8599 @AlidesChimin
Hi @RichardLitt ,
All requested modifications have now been implemented. The line-cleanup changes were merged (PR #7), the remaining formatting issues were resolved, and all bold text was removed as suggested. I also created a new release on GitHub and generated an updated Figshare archive for the JOSS submission:

https://doi.org/10.6084/m9.figshare.30776753

If you need any additional adjustments, I’ll be happy to update the package.
