### 2026-01-08 #8599 @AlidesChimin
> 🚨🚨🚨 **ISTO NÃO É UM SIMULADO, VOCÊ ACABOU DE ACEITAR UM TRABALHO NO JOSS!** 🚨🚨🚨
>
> Eis o que você deve fazer agora:
>
> 0. Confira o PDF final e os metadados do Crossref que foram depositados 👉 [Criando solicitação de pull para 10.21105.joss.08599  joss-papers#7616](https://github.com/openjournals/joss-papers/pull/7616)
> 1. Aguarde cinco minutos e verifique se o DOI do artigo resolve [em https://doi.org/10.21105/joss.08599](https://doi.org/10.21105/joss.08599)
> 2. Se tudo estiver correto, feche esta solicitação de revisão.
> 3. Comemore como se tivesse acabado de publicar um artigo! 🎉🌈🦄💃👻🤘
>
> Algum problema? Informe sua equipe técnica editorial...

I’ve checked the final PDF and metadata, and everything looks good.
