### 2026-01-09 #8599 @RichardLitt
@samhforbes Shouldn't this automatically close?
